+++ D3DFramework/D3DFramework/PathFinder.cpp
@@ -88,7 +88,7 @@ PKH::PathFinder::Node* PKH::PathFinder::FindStartNode(const Vector3& startPos, f
 		if (Physics::Raycast(ray, &hit, dist, mask)) continue;
 
 		// 기준반경 안에 있으면서, 사이에 벽이 없는 노드.
-		int id = dist * 100; // 정밀도를 위한 곱셈
+		int id = (int)(dist * 100); // 정밀도를 위한 곱셈
 		candidatesStart[id]= node.second;
 	}
 
@@ -123,7 +123,7 @@ PKH::PathFinder::Node* PKH::PathFinder::FindDestinationNode(const Vector3& destP
 		if (Physics::Raycast(ray, &hit, dist, mask)) continue;
 
 		// 기준반경 안에 있으면서, 사이에 벽이 없는 노드.
-		int id = dist * 100; // 정밀도를 위한 곱셈
+		int id = (int)(dist * 100); // 정밀도를 위한 곱셈
 		candidatesDest[id] = node.second;
 	}
